refactor(app): extract sync modules from useTarkov god file
Decompose useTarkov.ts (2017→1510 lines) into five focused modules:
- tarkov/syncTimeline.ts: self-origin detection for realtime updates
- tarkov/conflictDetection.ts: pure conflict detection between states
- tarkov/apiUpdateNotifier.ts: API update toast notification logic
- tarkov/realtimeListener.ts: Supabase realtime subscription
- tarkov/resetEngine.ts: reset/sync-pause utilities

Uses registerSyncControllerGetter pattern to avoid circular imports
between realtimeListener/resetEngine and the main store file.
